requirements.txt
tensorflow==2.12.0
h5py>=3.1.0,<4.0.0
numpy
matplotlib
opencv-python
pillow
streamlit